Fixed #5130 - Changed SessionId and PartitionKey filters on Azure Service Bus to check if the context is a service bus context.
