+++ src/Transports/MassTransit.Azure.ServiceBus.Core/AzureServiceBusTransport/Configuration/Topology/PartitionKeyMessageSendTopologyConvention.cs
@@ -1,6 +1,7 @@
 namespace MassTransit.AzureServiceBusTransport.Configuration
 {
     using MassTransit.Configuration;
+    using Topology;
 
 
     public class PartitionKeyMessageSendTopologyConvention<TMessage> :
